Added TU9 and Both GOTY Discs
Patches from illusion's RDR 0.0.0.1 2010 NTSC were ported to 2010 TU9, which is the recommended update of original game, and to both discs of the GOTY edition.

Original patch file renamed to match naming convention with new files.

"Bottomless Clip" exploit was patched out during a TU of original release, and couldn't be found in GOTY either.