Update, rename 45410914 - Rock Band 3 TU5.patch.toml
Added hash for new RB3 executables distributed as part of Rock Band 3 Deluxe. These were updated to fix an exploit in the game engine which could be used to execute arbitrary code (and was publicly exploited in the game Rock Band Blitz, which uses the same base engine, as an entry point for the BadUpdate exploit).
Also renames the TU5 patch file to use brackets for consistency with the majority of the patch set.